refactor(GUI): update the GUI to a single window
