+++ src/ramstk/views/gtk3/books/listbook.py
@@ -6,18 +6,20 @@
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """RAMSTK GTK3 List Book Module."""
 
+# Standard Library Imports
+from typing import List
+
 # Third Party Imports
 from pubsub import pub
 
 # RAMSTK Package Imports
 from ramstk.configuration import RAMSTKUserConfiguration
 from ramstk.logger import RAMSTKLogManager
-from ramstk.views.gtk3 import _
+from ramstk.views.gtk3 import GObject, Gtk
 from ramstk.views.gtk3.revision import lvwFailureDefinition, lvwUsageProfile
-from ramstk.views.gtk3.widgets.basebook import RAMSTKBook
 
 
-class RAMSTKListBook(RAMSTKBook):
+class RAMSTKListBook(Gtk.Notebook):
     """
     This is the List Book class for the GTK3 multiple window interface.
 
@@ -30,6 +32,14 @@ class RAMSTKListBook(RAMSTKBook):
         Key is the RAMSTK module name; value is a list of Views associated with
         that RAMSTK module.
     """
+
+    dic_tab_position = {
+        'left': Gtk.PositionType.LEFT,
+        'right': Gtk.PositionType.RIGHT,
+        'top': Gtk.PositionType.TOP,
+        'bottom': Gtk.PositionType.BOTTOM
+    }
+
     def __init__(self, configuration: RAMSTKUserConfiguration,
                  logger: RAMSTKLogManager) -> None:
         """
@@ -40,8 +50,8 @@ class RAMSTKListBook(RAMSTKBook):
         :param logger: the RAMSTKLogManager class instance.
         :type logger: :class:`ramstk.logger.RAMSTKLogManager`
         """
-        RAMSTKBook.__init__(self, configuration)
-        self.dic_books['listbook'] = self
+        GObject.GObject.__init__(self)  # pylint: disable=non-parent-init-called
+        self.RAMSTK_USER_CONFIGURATION = configuration
 
         # Initialize private dictionary attributes.
         self._dic_list_views = {
@@ -67,6 +77,7 @@ class RAMSTKListBook(RAMSTKBook):
         }
 
         # Initialize private list attributes.
+        self._lst_handler_id: List[int] = []
 
         # Initialize private scalar attributes.
 
@@ -95,15 +106,7 @@ class RAMSTKListBook(RAMSTKBook):
                 )]
         except KeyError:
             _tab_position = self._bottom_tab
-        self.notebook.set_tab_pos(_tab_position)
-
-        self.set_title(_("RAMSTK Lists and Matrices"))
-        self.set_deletable(False)
-        self.set_skip_pager_hint(True)
-        self.set_skip_taskbar_hint(True)
-
-        self.resize((self._width / 3) - 10, (2 * self._height / 7))
-        self.move((2 * self._width / 3), 0)
+        self.set_tab_pos(_tab_position)
 
     def _on_close(self) -> None:
         """
@@ -133,7 +136,5 @@ class RAMSTKListBook(RAMSTKBook):
         :return: None
         :rtype: None
         """
-        RAMSTKBook.on_module_change(self)
-
         for _list in self._dic_list_views[module]:
-            self.notebook.insert_page(_list, _list.hbx_tab_label, -1)
+            self.insert_page(_list, _list.hbx_tab_label, -1)
