+++ src/ramstk/views/gtk3/books/workbook.py
@@ -6,19 +6,29 @@
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """RAMSTK GTK3 Work Book Module."""
 
+# Standard Library Imports
+from typing import List
+
 # Third Party Imports
 from pubsub import pub
 
 # RAMSTK Package Imports
 from ramstk.configuration import RAMSTKUserConfiguration
 from ramstk.logger import RAMSTKLogManager
-from ramstk.views.gtk3 import _
+from ramstk.views.gtk3 import GObject, Gtk
 from ramstk.views.gtk3.revision import wvwRevisionGD
-from ramstk.views.gtk3.widgets.basebook import RAMSTKBook
 
 
-class RAMSTKWorkBook(RAMSTKBook):
+class RAMSTKWorkBook(Gtk.Notebook):
     """This is the Work Book for the pyGTK multiple window interface."""
+
+    dic_tab_position = {
+        'left': Gtk.PositionType.LEFT,
+        'right': Gtk.PositionType.RIGHT,
+        'top': Gtk.PositionType.TOP,
+        'bottom': Gtk.PositionType.BOTTOM
+    }
+
     def __init__(self, configuration: RAMSTKUserConfiguration,
                  logger: RAMSTKLogManager) -> None:
         """
@@ -29,8 +39,8 @@ class RAMSTKWorkBook(RAMSTKBook):
         :param logger: the RAMSTKLogManager class instance.
         :type logger: :class:`ramstk.logger.RAMSTKLogManager`
         """
-        RAMSTKBook.__init__(self, configuration)
-        self.dic_books['workbook'] = self
+        GObject.GObject.__init__(self)  # pylint: disable=non-parent-init-called
+        self.RAMSTK_USER_CONFIGURATION = configuration
 
         # Initialize private dictionary attributes.
         self._dic_work_views = {
@@ -63,6 +73,7 @@ class RAMSTKWorkBook(RAMSTKBook):
         }
 
         # Initialize private list attributes.
+        self._lst_handler_id: List[int] = []
 
         # Initialize private scalar attributes.
 
@@ -90,18 +101,7 @@ class RAMSTKWorkBook(RAMSTKBook):
                 )]
         except KeyError:
             _tab_position = self._bottom_tab
-        self.notebook.set_tab_pos(_tab_position)
-
-        self.set_title(_("RAMSTK Work Book"))
-        self.set_deletable(False)
-        self.set_skip_pager_hint(True)
-        self.set_skip_taskbar_hint(True)
-
-        _width = self._width - 20
-        _height = (5 * self._height / 8) - 40
-
-        self.resize(_width, _height)
-        self.move(1, (_height / 2) + 100)
+        self.set_tab_pos(_tab_position)
 
     def _on_module_change(self, module: str = '') -> None:
         """
@@ -110,7 +110,5 @@ class RAMSTKWorkBook(RAMSTKBook):
         :return: None
         :rtype: None
         """
-        RAMSTKBook.on_module_change(self)
-
         for _workspace in self._dic_work_views[module]:
-            self.notebook.insert_page(_workspace, _workspace.hbx_tab_label, -1)
+            self.insert_page(_workspace, _workspace.hbx_tab_label, -1)
